requirements.txt
Flask==2.3.3
Whoosh==2.7.4
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
pyspellchecker==0.7.2
python-dotenv==1.0.0
email-validator==2.0.0
pytz==2023.3 